Command Stack Working

+++ src/Strategy.java
@@ -1,5 +1,5 @@
 public interface Strategy {
     public void input();
     public String parse_expression(String expression);
-    public int evaluate_expression(String parse);
+    public float evaluate_expression(String parse);
 }
